backend/requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
setuptools>=65.0.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9

# Cache & Task Queue
redis==5.0.1
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.6.0

# File Storage
boto3==1.34.34
django-storages==1.14.2
minio==7.2.0

# Configuration
python-decouple==3.8

# Authentication
djangorestframework-simplejwt==5.3.0
django-allauth==0.61.1

# Payments (Stripe)
stripe==7.14.0

# Image processing
# Pillow==9.5.0  # Python 3.13 호환성 문제로 임시 제거

# WSGI Server
gunicorn==21.2.0

# Development
django-debug-toolbar==4.2.0